Minor fixes for analytics submission
