+++ lib/submit-analytics.js
@@ -219,7 +219,7 @@ async function submitAnalytics(requestFactory, context, eventName) {
         
         // on v6.x, we submit both MixPanel and S3 analytics
         // MixPanel analytics make no distinction between production and staging modes
-        versionTags[0] < 10 ? dispatchAnalytics(new MixpanelDetails(), payload) : null,
+        versionTags[0] < 10 ? dispatchAnalytics(MixpanelDetails, payload) : null,
     ]);
 }
